Use new MusicPlayer interface for pitch, speed and position sliders
